PageDisplay FC - logic from SignUpPage
